requirements.txt
streamlit
python-dotenv 
openai
ffmpeg-python
#gemini api
google-generativeai


# #realtime api re:
# pyaudio
# websockets